LLMCaller/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
attrs==24.2.0
certifi==2024.8.30
distro==1.9.0
exceptiongroup==1.2.2
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pydantic==2.9.2
pydantic_core==2.23.4
referencing==0.35.1
rpds-py==0.20.0
sniffio==1.3.1
typing_extensions==4.12.2
